fix: patches for building on modern Linux with GCC 15 and NixOS
- html.h: fix K&R-style declaration (unspecified args rejected by GCC 15/C23)
- sys_info.cpp: guard sys/sysctl.h behind #ifndef __linux__ (removed in glibc 2.32)
- m_backtrace.cpp: fix size_t vs unsigned mismatch on 64-bit
- configure: fix missing ] in fontdir handling, clear problematic GCC version list,
  lower minimum GCC version requirement, replace FindDir checks that fail in
  sandboxed builds (no /usr/include)

+++ src/mstl/m_backtrace.cpp
@@ -49,7 +49,7 @@ backtrace::~backtrace() throw() {}
 
 backtrace const& backtrace::operator=(backtrace const&) { return *this; }
 bool backtrace::is_debug_mode() { return false; }
-void backtrace::text_write(ostringstream&, size_t) const {}
+void backtrace::text_write(ostringstream&, unsigned) const {}
 
 #else // __OPTIMIZE__
